remove occupancy widget and other improvements
