+++ crates/rmf_site_editor/src/lib.rs
@@ -63,7 +63,7 @@ use site::{OSMViewPlugin, SitePlugin};
 use site_asset_io::SiteAssetIoPlugin;
 
 pub mod mapf_rse;
-use mapf_rse::MapfRsePlugin;
+use mapf_rse::NegotiationPlugin;
 
 pub mod osm_slippy_map;
 use bevy::render::{
@@ -290,7 +290,7 @@ impl Plugin for SiteEditor {
             app.add_plugins((StandardUiPlugin::default(), MainMenuPlugin))
                 // Note order matters, plugins that edit the menus must be initialized after the UI
                 .add_plugins((site::ViewMenuPlugin, OSMViewPlugin, SiteWireframePlugin))
-                .add_plugins(MapfRsePlugin::default());
+                .add_plugins(NegotiationPlugin::default());
         }
 
         if self.is_headless_export() {
